Rename the fluxion.core.perception package
