+++ src/fluxion/perception/sources/audio_sources.py
@@ -1,6 +1,6 @@
 from abc import abstractmethod
 import numpy as np
-from fluxion.core.perception.sources.perception_source import PerceptionSource
+from fluxion.perception.sources.perception_source import PerceptionSource
 import sounddevice as sd
 
 class AudioSource(PerceptionSource):
